feat : 검증 코드 추가, 출력 입력 코드 추가, 바우처 타입 enum 코드 추가, yaml 설정 변경 기존 서비스에 너무 많은 코드가 있어서 나눴습니다.

+++ src/main/java/org/prgms/management/entity/PercentAmountVoucher.java
@@ -4,11 +4,11 @@ import java.util.UUID;
 
 public class PercentAmountVoucher implements Voucher {
     private final UUID voucherId;
-    private final long percent;
+    private final int percent;
     private final String voucherName;
     private final String voucherType;
 
-    public PercentAmountVoucher(UUID voucherId, long percent, String voucherName, String voucherType) {
+    public PercentAmountVoucher(UUID voucherId, int percent, String voucherName, String voucherType) {
         this.voucherId = voucherId;
         this.percent = percent;
         this.voucherName = voucherName;
